rewrite: refactor AbstractInterfaceView & InterfaceListeners
removes combined / player interfaces, unsure if I want to readd them
later. reactivity works and the overhead of async/deferred transforms
has been reduced with virtual threads. it needs some testing on what
happens if a transfrom runs while the interface is trying to close but
it mostly works.

todo:
  - revisit method names
  - simplify elements
  - properly implement back button
  - add some kind of handling for clicks in the user part of the
    inventory